13.1. norminette

+++ Libft/ft_memset.c
@@ -6,13 +6,15 @@
 /*   By: thbui <thbui@student.42prague.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/12 14:13:48 by thbui             #+#    #+#             */
-/*   Updated: 2023/01/13 15:45:22 by thbui            ###   ########.fr       */
+/*   Updated: 2023/01/13 16:25:28 by thbui            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "libft.h"
+
 void	*ft_memset(void *s, int c, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (n > i)
